Make quick_sorter work with move-only types (issue #15).
Until there, quick_sorter didn't work with move-only types because it
needed to copy the copy before the partitioning operations. The new
implementation moves the pivot out of the partitioned area and moves it
back in its final position. Moreover, the pivot selection was changed to
use a median of 9 pivot selection; otherwise, the push_front pattern
triggered a worst-case behaviour.