Make sorter_traits detect embedded traits.
The new sorter_traits is implemented in such a way that its generic
version defines traits only if they exist in the passed sorter. The
rationale is that one doesn't always need every trait and thus we can make
efforts to reduce the coupling between traits as long as it has no cost
for users.