Add variadic trait checkers.
The new checkers are split into to categories: the raw trait checkers,
that look directly whether a type contains the appropriate type alias
(those are used by sorter_traits), and the high-level ones, that use
sorter_traits to check whether the required type alias exists. The
high-level type checkers are meant to be used by sorter adapters.
Moreover, the checkers are now variadic so that they can aggregate the
iterator categories and stability of a group of sorters into the most
restrictive combination of them.