Make merge_insertion_sort use less memory.
To know which node it had to insert in the main chain next,
merge_insertion_sort stored the indices of the pend nodes and searched the
node whose index corresponded to the wanted Jacobsthal number. The new
implementation doesn't store the pend indices anymore. Instead, when we
need to find the next pend element to insert into the main chain, it
computes the difference between the wanted Jacobsthal number and its
predecessor and picks the pend element which is that far from the
beginning of the list. For performance matters, we store the difference
between adjacent Jacobsthal numbers instead of the number themselves.