Directly put first pend element in main chain.
In a merge-insertion sort, the first pend element is always the lowest
element of the main chain. Until there, the algorithm initialized the main
chain and the pend elements list, then it moved the first pend element at
the beginning of the main chain. This commit puts the first pend elemet
directly into the main chain instead, saving a bit of work when there are
few elements. The condition at the beginning of the algorithm ensures that
there will always be at least two elements to sort when the chain list is
initialized.