Fix merge_insertion_sort.
The number of comparisons performed by merge_insertion_sort was too high.
The reason is that the iterator needed to find the next pend element to
insert corresponding to a Jacobsthal number was advanced one too many
times. Instead of advancing the the iterator dist - 1 times, we subtract 1
to every element of the jacobsthal_diff array. It should be slightly
faster.