More memory-efficient algorithm.
The merge-insertion sort doesn't store the original iterators in the list
anymore but uses the fact that the position of the pend iterator in the
range [first, last) is known: we just have to iterate the collections two
elements at a time to find them. Also, it doesn't erase the pend elements
from the list anymore but lets the destructor handle that. Also, since
constant removal from the beginning was not needed anymore, a vector is
used to store the pend elements instead of a list (which only stores the
upper bounds of insertion in the main chain).