Make small_array_adapter use fixed_sorter_traits.
The fixed_sorter_traits class template now contains information about the
specializations of N it works with; it's called the domain.
small_array_adapter has been updated so that it can take advantage of this
information when it exists, and default to working with any size when the
information does not exist.