Remove std::greater<T> overloads in string_spread_sorter.
During the development of the library, I made a conscious choice to drop
support for the old function objects and to only use the enhanced C++14
ones everywhere. To enfore that choice, this commit removes the special
overload taking std::greater<std::string> and std::greater<std::wstring>
in string_spread_sorter. The goal is to consistently force users to use
std::less<> and std::greater<>.