Remove has_sort_method from utility.
The metafunction has_sort_method was only used by self_sort_adapter and
has few obvious uses outside of this sorter adapter. Every other similar
metafunction in the library is part of the details, so there was no reason
for this function to be exposed. Therefore, this commit moves it from
utility to the details of self_sort_adapter.