Improve measure of presortedness: inv.
To achieve O(n log n) complexity, inv needs to allocate memory, which
means that it can fail when not enough memory is available. This commit
changes inv so that it is O(n log n) when enough memory is available and
falls back to a O(n^2) algorithm when there isn't enough memory.