Remove std::less<T> and std::greater<T> swap_if.
The decision was made a while ago that cpp-sort would only try to support
the C++14 transparent function objects and drop support for std::less<T>
and std::greater<T>. It seems that swap_if had some remnants of the time
when the old functions objects were supported. This commit removes them.