Introduce iter_swap_if.
iter_swap_if is equivalent to swap_if, except it takes a pair of iterators
instead of a pair of values and swaps the dereferenced iterators.
Actually, iter_swap_if will use an ADL-found iter_swap if there is one, or
utility::iter_swap if it finds a suitable iter_move overload with ADL. If
it doesn't need to use an ADL-found iter_swap or iter_move, it will call
swap_if and might benefit from its optimizations.