Proxy iterators support for fixed-size sorters.
Replace almost every use of swap_if in the library by an equivalent call
to the new iter_swap_if; it mostly concerns fixed-size sorters. Also
update generate.py to generate calls to iter_swap_if instead of swap_if.
iter_swap_if is not used as a customization point yet but if could if the
need arises.