Fix bidirectional unbuffered merge_sort.
The merge_sort optimization used to save memory when the collection is
already mostly sorted actually broke merge_sort for bidirectional
iterators when no more memory was available since it failed to decrease
the total length of the collection when the size of the left range
decreased. This commit fixes the bug by computing again the size of the
right range from the original size alone rather than from the size of the
left range.