Optimized schwartz_adapter with no projection.
When it isn't given any projection, schwartz_adapter currently creates a
copy of every element projected with utility::identity, which is a total
waste of space and time. This commit alters schwartz_adapter so that it
forwards its parameters directly to the wrapped sorter when it isn't given
any projection.