Add a stable_adapter to the library.
This commit introduces a stable_adapter template, which takes a sorter and
makes sure that it takes into account the original position of equivalent
elements when it sorts the collection. To do so, it stores the original
position of every element and compares it to decide which element should
go first when two elements are equivalent. This version is suboptimal, but
improvements will be added to optimize it when possible.