Fix proxy iterators for inplace_merge.
Add a way to create assign an association&& instance to an association.
Use it to fix the implementation of inplace_merge. It fixes a hole for the
case when ones needs to move an association object to another location in
the same collection, and eventually makes verge_sorter work smoothly with
proxy iterators.