Forward declarations for the library.
It is sometimes useful to have forward declarations of the main library
components, for example to provide explicit specializations of adapters
for some sorters, or to forbid specific sorters from being used. This
commit introduces the header fwd.h which contains forward declarations for
every sorter, fixed-size sorter, and sorter adapter in the library. It
also makes default_sorter and spread_sorter structs so that they can be
forward-declared like the other sorters.