Expose the underlying mechanism of stable_adapter.
stable_adapter actually short-circuits the generic mechanism when the
wrapped sorter is already stable, and is meant to be explicitly
overridable. People might want to access the underlying mechanism to make
a sorter stable. This commit introduces make_stable, a class template that
implements the underlying mechanism of the generic stable_adapter, but
which isn't meant to be explicitly specialized whatsoever.