Rollback a few changes.
The enhancement to buffered sorters doesn't actually work as expected
since it's impossible to pass explicit template parameters to operator()
with the regular call syntax, we have to do it with the .operator()<T>()
syntax which kind of defeats the purpose, so we roll the change back. Also
the deque in the testsuite is changed to a vector to avoid Valgrind
complaining about timsort. An issue has been submitted to the original
timsort implementation.