Human-readable poplar_sorter's sift function.
Most of the functions in poplar_sorter were originally implemented by
copying the implementation exposed in Smoothsort Revisited, then gradually
modified to suit the new needs, then to fix a Valgind error. Even though
it worked, the function sift became unreadable. I rewrote it to be a bit
more human-friendly. Performance is the same.