Make small_array_adapter even more generic.
A previous update made small_array_adapter accept more fixed-size sorters,
so that it can also adapt those that have additional type template
parameters. However, it failed to forward those parameters to the actual
fixed-size sorter. This commit fixes that. Note however that things might
go awry if the underlying fixed-size sorter is supposed to accept an
std::index_sequence as a parameter.