Move spreadsort type constraints to the sorter.
Deep down in the implement, integer_spread_sort checked that the integers
to be sorted had a size that was both smaller than std::size_t and smaller
than std::uintmax_t, and called pdqsort if it wasn't the case. This commit
moves this check to the sorter to benefit from the top-level SFINAE. It
also cleans the remnants of the corresponding changes made in
float_spread_sorter a while ago.