Buffers need not expose pointers.
The buffers used by buffered sorters were required to return pointers and
to have a data() member function. This commit removes the data() overloads
in fixed_buffer and dynamic_buffer and alters grail_sort and block_sort so
that they can use the returned buffer as long as it exposes random-access
iterators.