Refactor is_sorter_* traits (even more).
Sorter traits were usable as variable templates. However, future
evolutions have shown that we also need them as structs similar to those
in the standard library. This commit changes sorter traits to be structs
instead of simple variable templates. To keep things relatively simple, it
also adds *_v variable templates to be drop-in replacements for the old
traits.