Test sorters with move-only types.
Add test to make sure that every sorter is able to handle move-only types.
The test type is also enhanced so that it can check whether a
read-after-move occured. The buffered sorters only use a fixed buffer of
size 0 because the current sorters are not meant to handle move-only
types.