Revamp destruct_n.
The class destruct_n was taken as is taken from libc++, but I didn't like
everything about it, like how verbose it was to call any function. This
commit moves the template information to the class (instead of every
member function). It also allows not to store the size when the type is
trivially destructible.