Mergesort for lists.
Add a specialization of container_aware_adapter for merge_sorter that
recognizes std::forward_list and std::list, and uses splicing instead of
swaps to insert the elements without invalidating the iterators. This
commit also fixes a problem in the mechanism used to conditionally include
headers that contain specializations of library templates for other
library types.