Early return for vergesort when the data is sorted.
One of the first functions called by vergesort is is_sorted_until since it
easily allows to find the length of the first ascending run. This commit
uses the information to check whether the collection is already sorted
with a cheap iterator comparison and makes vergesort return early if it is
the case.