self_sort_adapter now knows standard lists.
std::forward_list and std::list have sort methods implementing a stable
sort. To be honest, those are the classes that are the most likely to be
used together with self_sort_adapter, so the adapter was updated to take
into account the fact that these lists' sort method is stable even though
it isn't named stable_sort.