Add refined case_insensitive_less
The generic version of case_insensitive_less is slow, especially for small
strings compared to a simple Schwartz transform with std::tolower. The
refined version calls std::use_facet only when created while the generic
version calls it whenever the comparison is performed, leading to dramatic
loss in performance for almost nothing.