Remove ska_sort_copy, fallback to pdq_sort
Currently, cpp-sort only offers an in-place sorting interface. There are
plans to work on a sort_copy interface, but not right now (issue #42),
hence I removed everything related to ska_sort_copy.

The fallback of ska_sort was change to pdq_sort instead of std::sort to
offer a better fallback algorithm, and to ensure that proxy iterators work
as expected with ska_sort.