Slightly improve selection_sort
Selection sort is a repeated call to min_element. However, we know that
min_element won't ever be given an empty range with selection_sort, hence
the addition and use of a new unchecked_min_element which does the same
thing a min_element, except that it doesn't check whether the range is
empty.