Fix bug in minmax_element_and_is_sorted (fixes #103)
When a collection was already sorted, minmax_element_and_is_sorted
returned the one-passed-the-end iterator for 'max' instead of the end
iterator, leading to undefined behaviour. It now returns the last
dereferenceable iterator as it should have done.

This notably fixes a bug in counting_sorter that was quite hard to track
because of the undefined behaviour.