Remove custom begin and end utility functions
I introduced these functions a while ago when trying to handle span-like
classes, but totally forgot back then that Iterable&& bound to an rvalue
but was itself an lvalue, obviating the need for begin and end functions
working with rvalues.

Actually, there doesn't seem to be any safe use case for begin and end
functions taking rvalues, so we're better off using std::begin and
std::end and getting rid of the custom functions. On a side note, Range-v3
and the Ranges TS mark the rvalue overloads as deprecated, but kept for
backward compatibility.