New timsort implementation from tvanslyke/timsort-cpp
The new implementation seems to be a bit slower than timsort-cpp itself, I
will have to investigate to find what I did wrong. On the other hand it
passes the whole testsuite, which is great.

I still need to clean it and factor out things that already exist in
cpp-sort.

If we're lucky it also solves the std::deque bug that I had with the gfx
version of timsort.