Add utility adapter_storage
We want to be able to adapt mutable sorters, but also to still transform
the adapted sorter to a function pointer when the sorter is not mutable.
Hence we introduce utility::adapter_storage stores the passed sorter only
when std::is_empty is std::false_type, allowing an adapter to be
conditionally empty depending on the sorter to adapt.