Fix a "bug" in poplar sort
When the size of the collection to sort with poplar sort was already a
number k such as k = 2^n-1, poplar sort made three poplars out of it
instead of one. For example for a collection of size 63, it made poplars
with sizes 31, 31 and 1, which isnt as efficient as it could, but also
goes against the definition of the algorithm which says that only the last
two poplars can have the same size.