Don't use sorter_facade to implement hybrid_adapter
The operator() in hybrid_adapter should already be enough to correctly
call the right operator() overload for the underlying sorters since said
sorter are already supposed to provide the same overload as if they had
been implemented with sorter_facade.

This once again reduces slightly the size of the produced binary.