Remove deduction guides for scope_exit
We actually checked that std::uncaught_exceptions was available for
scope_exit, but used deduction guides which are not always available
when std::uncaught_exceptions is (case in point: -std=gnu++14). This
commit removes the deduction guide and uses a make_* function instead
so that scope_exit-dependent code works more often.