Reduce usage of cppsort::sort in the test suite (issue #167)
cppsort::sort and cppsort::stable_sort are deprecated and will be
removed in version 2.0.0. This commit drastically reduces their use in
the test suite and only keep them where they are specifically tested
because they're relevant.

This should also reduce the parsing time of the test suite: since the
header drags default_sorter along the way it tends to be heavy for tests
where a single sorter is needed.

This commit also fixes a few minor style issues or include issues that
were hidden by transitive includes along the way.