Improve adapter tests for forward and bidirectional iterators
Add specific tests to tests that sorters wrapped with schwartz_adapter
and stable_adapter can properly sort collections that expose forward
iterators or bidirectional iterators.

As a drive-by change, several std::begin(xxx)/std::end(xxx) were turned
into xxx.begin()/xxx.end() as part of a long-running change to avoid
calls to overly generic helpers in non-generic code. The goal is to
improve error messages and reduce the amount of overload resolution the
compiler has to perform. I don't plan to apply this change to the whole
test suite in one go, but to apply it on a case-by-case basis when the
test suite files are changed for various other reasons.