verge_sort: make better use of existing size information (#169)
Add new innplace_merge overload for bidirectional iterators to allow to
pass the sizes of the subranges to merge without passing a buffer. Use
this information to avoid recomputing size information as often in the
vergesort overload that handles bidirectional iterators.

This commit also cleans the bidirectional overload of vergesort, which
had been kind of neglected.