Fix the iterator_category of out_of_place_adapter
No matter the sorter it adapts, out_of_place_adapter performs an
indirect sort by sorting elements out-of-place in a contiguous memory
buffer, so random-access sorters can be used even if the iterators
passed to the resulting sorter are forward iterators. This commit
actually gives the std::forward_iterator_tag iterator_category to
out_of_place_adapter in order to match its abilities and the
documentation.

This commit adds tests which also happen to test that indirect_adapter
follows the same convention.