Embed wiki sources in a 'docs' subdirectory (issue #157)
The new subdirectory allows contributers to propose pull requests to
complete or fix the documentation, and also allows to enforce the good
practice of modifying the documentation along the code, making it easier
to track the changes to revert in the documentation if needed.

The new subdirectory was named "docs" to follow the Pitchfork project
layout conventions (https://github.com/vector-of-bool/pitchfork).

[ci skip]