Update benchmarks
The benchmarks wiki page was rewritten, which was long overdue since the
graphs and comments were mostly old and did not correspond to the
current algorithms but to verious epochs in the development of the
library. The new page uses all the latest improvements to algorithms and
benchmarks, and the comment were updated accordingly. Moreover, I tried
to restructure the whole thing better.

Thanks Mikhail for fixing typos.

[ci skip]