Add a new benchmark for Inv-adaptive algorithms
This benchmark reuses code from other benchmarks, so some functions were
split to common utility files to limit the amount of copy-pasted code.
This commit also changes the "inversions" distribution to actually be
useful, and uses it in the new benchmark.

The Benchmarks.md page of the documentation was also updated with the
results of the benchmark against drop_merge_sort and split_sort.

[ci skip]