Fix big in median-of-9 pivot selection
Median-of-9 pivot selection takes 9 samples in a collection, takes the
medians of the 3 sets of 3 samples, then takes the median of those three
medians. Our implementation incorrectly picked two correct medians and
a value that wasn't a median for the last triplet of samples. This is
now fixed with this commit.

How it affects the speed of the quicksort family of algorithms is
unclear. It might make a difference for specific patterns.