Revamp the bidirectional version of vergesort
The bidirectional version of vergesort was modified as follows:
- It now uses the same k-way merge as the random-access version.
- It creates fewer partitions with one element, generally being smarter
  about which run an element should belong to.
- It spends less time recomputing runs sizes, leading to a lower impact
  on the running time overall.

The k-way merge algorithm is now shared between both versions and has
been tweaked to handle 1-element runs better. The in-code documentation
was also reworked to be more comprehensible and comprehensive. Some
variable names were changed to make the logic more obvious.