New feature: cppsort::stable_t
The type alias stable_t is kind of a better equivalent to
stable_adapter: it aliases either a sorter directly if it is always
stable, otherwise stable_adapter<Sorter>::type if it exists, otherwise
stable_adapter<Sorter> itself. The goal is to alias the "least nested"
type that is always stable to decrease template instantiation, template
depth, and somehow bonary size. It also improves the error messages is
some cases.

Bundled with this feature: stable_adapter specializations now expose a
member type named 'type' to implement the mechanism described above.

The documentation page about sorters adapters was slightly cleaned up
and reworked as part of this change.