Add an explicit stable_adapter<verge_sorter> specializations
An likewise for stable_adapter<verge_adapter>. Instead of simply relying
on make_stable, the new specializations use a variant of vergesort that
detects strictly descending runs instead of non-ascending ones, and
wraps the fallback sorter in stable_t.

The technique of detecting non-descending and strictly descending runs
is a trick borrowed from timsort in order to preserve stability in
natural mergesort algorithms: descending runs are reversed in-place, so
equivalent elements wouldn't retain their original order when reversed
in such runs.

While verge_adapter does not handle bidirectional iterators, the
underlying code was changed so that it will be easier to do in the
future, but currently this involves some dirty tricks with a stripped
down implementation of a sized iterator. This will make the transition
to C++20 ranges and standard sized iterators and sentinels easier in the
long term. Meanwhile it is just an implementation detail.

This commit addresses external issue Morwenn/vergesort#11 and somehow
Morwenn/vergesort#7 too. cpp-sort is a better recipient for those
variations on vergesort than the standalone project, but some
cross-project documentation will be needed anyway.