Categorize swap_ranges invocations (#180)
Split swap_ranges calls into swap_ranges_inner and swap_ranges_overlap,
where the first has the "no overlap" guarantee and checks to ensure that
said guarantee holds.

This commit also adds an audit mechanism to cpp-sort for checks that
are too expensive for CPPSORT_ASSERT, and starts using it in
swap_ranges_inner to ensure that the "no overlap" guarantee holds.