Make sorter conversion to function pointer unconditionally constexpr
That feature used to only be available in C++17, when the conversion of
lambdas to function pointers was made constepxr. The new design doesn't
use lambdas anymore, which allows the conversion to always be constexpr.

This commit also simplifies the implementation, making it less redundant
and easier to extend. Not using lambdas also means easier to understand
error messages and less potential generic lambda bloat.