Add measure of presortedness: sus (issue #62)
Add the measure of presortedness SUS(X) first described in Sorting
Shuffled Monotone Sequences by Levcopoulos and Petersson.

SUS(X) is the minimum number of non-decreasing subsequences (of possibly
non adjacent elements) into which X can be partitioned. It happens to be
equal to the size of the longest decreasing subsequence of X minus 1,
which is why we reuse the algorithm used to compute Rem(X) to compute
it.