scope_success::release -> activate/deactivate
I will soon need to deactivate and reactivate scope guards, so this
commit introduces the member function activate() and changes release()
into deactivate(). I wanted to keep release() to keep parity with the
standard scope guards, but acquire() felt more obscure and not obvious
enough, so activate/deactivate it is.