Reduce the allocations in merge_insertion_sort
Every recursion of merge_insertion_sort_impl used to allocate its own
list node pools, leading to 2N node allocations. This commit changes it
so that it only allocates N nodes once and passes the node buffer down
the recursion to reuse nodes from the same pool.

However that led to a big speed degradation, most likely due to the fact
that the subsequent accesses to the list nodes were way less
cache-friendly. Therefore, this commit also adds invariants-breaking
black magic to the node pool so that the node links can be reset between
node pool uses. This effectively allows to have the same performance
than before while decreasing the number of allocations.