Bring more uniformity in indirect algorithms
Algorithms that needed to perform indirect operations (operations on a
collection of iterators instead of on a collection of values) mostly
used indirect_compare so far, which embedded both the comparison and the
projection. This commit introduces indirect() to only adapt the
projection, making it it simpler and more powerful. This new function
object also allows to replace the lambdas used for that exact purpose in
indirect_adapter, bringing more uniformity into the library.