Improve melsort encroaching list construction
Instead of performing a binary search on the encroaching list that
checks both the greatest and smallest element, compare first to the head
and tail of the last encroaching list, then decide whether to binary
search the heads, binary search the tails, or create a new list where to
add the element.

This optimization was actually described in the original paper, but I
failed to understand it until recently. It is possible because the last
list has the biggest of the heads (smaller elements), and the smallest
of the tails (bigger elements). At worst the number of comparisons is
the same as with the previous algorithm, at best it performs up to 20%
fewer comparisons.

I haven't noticed a big difference when sorting arithmetic types, but it
should make a noticeable difference for types with a more expensive
comparisons.