Reduce the number of comparisons performed in probe::enc
This is basically the optimization that was implemented in melsort a few
commits ago: it compares every new element to the head and tail of the
last encroaching list first, then decides whether to perform a binary
search over the heads or the tails of the encroaching lists. This
optimization leads to up to 40% fewer comparisons performed in some
cases.