Template fixed_size_list over the node type
This will allow future algorithms to use different but compatible node
types as needed. The node pool was changed in a similar way: at some
point it is just a pool of nodes that share some properties, allowing
more types of nodes is good.

fixed_size_list_node was renamed to list_node because it could very well
be used in non-fixed-size lists in the future.