New sorter: slab_sorter
Implements the slabsort algorithm described by C. Levcopoulos and O.
Petersson in Sorting Shuffled Monotone Sequences.

Despite its suspiciously simple-looking 7 lines of pseudocode, this
algorithm required non-trivial changes in fixed_size_list and a small
refactoring of melsort, as weel as the addition of stable_partition to
the library details.