Optimize probe::enc with lower_monobound_n when possible
Monobound binary search is a variation on standard binary search
proposed by @scandum: instead of computing the exact size of the next
partition where to look for an element, it always recurses in a
partition of size n-n/2. This change means that a few redudant
operations are performed when the size isn't a power of 2, but it also
means that the generated code can be much smaller.

Here we use it for probe::enc when the comparison and projection
functions are likely branchless: benchmarks showed that tis measure of
presortedness becomes up to 40% when measuring the presortedness of a
std::vector<double>. The gain is sufficient to switch to monobound
binary search for the cases where it is an obvious improvement.
probe::enc still uses a standard binary search the rest of the time.