Improve container_aware_adapter(insertion_sort) over std::list
The algorithm was using upper_bound in the inner loop, which was
recomputing the size of the range where to insert the new element at
each iteration. This changes the algorithm which now manages the size
itself and feeds it to upper_bound_n instead, saving O(n) operations per
iteration.