fixed_size_list: extract iteration to list_node_base
Many operations on list nodes only act on the iteration and the
relinking, and the specific type of the list node isn't required for
those. This creates a new non-template list_node_base class with the
pointers prev & next, and changes many operations to accept a pointer to
a list_node_base instead of a node_list<T>.

The major gain is that it allows to use a list_node_base for sentinel
nodes instead of a full list_node<T>, saving sizeof(T) space for every
list - space that was never used anyway. This means that the memory cost
of melsort is reduced by up to sqrt(n)*sizeof(T) while that of slbasort
is reduced by up to 2log(n)*sizeof(T).