Introduce immovable_vector class
Really similar to std::sort, but with the following properties: it's got
a fixed capacity determined at construction time, and it is not movable.
Unlike std::vector it allows to store immovable types.

We use that new class everywhere it makes sense, simplifying a bunch of
std::unique_ptr/destroy_n thingies.

This commit also changes cartesian_tree_sort quite a bit and fixes a
potential bug in its destructor: when an exception was thrown during the
construction of the Cartesian tree, destroy_at was called on every
allocated element instead of every constructed element.