Use a O(n) algorithm for probe::par (#85)
The old algorithm was an unsightly O(n² log n) algorithm, and a really
cunning on at that since it returned in the blink of an eye (generally
even faster than that) for most inputs, and only hit its worst case for
a few inputs.

The new algorithm is a O(n) algorithm described by T. Altman and Y.
Igarashi in *Roughly Sorting: Sequential And Parallel Approach*. It runs
in O(n) space and doesn't run as fast for most inputs, but it at least
doesn't hide as many surprises.

The old algorithm is still used when not enough memory is available, to
ensure that the next 1.x.y release won't be a breaking change, but it
will be removed in cpp-sort 2.0.0.