Micro-optimization for probe::dis
The value of Osc(X) is obtained by computing j - i, and can only grow
through the course of the algorithm. Since i always decreases, if it
becomes smaller or equal to the current value of Osc(X), then it means
that we're done because we can't find a bigger Osc(X) than the current
value.