Fix segfault on OSX
The segfault was introduced with the previous commit, and most likely
linked to storing a pointer to a thread_local object and trying to
access it from another thread.

This commit shuffles things around to avoid that issue, but also to
avoid random thread_local issues with MinGW-w64 which don't feel
deserved unlike the OSX one.