Make sorter_traits<stable_adapter<T>> always stable
This is mostly to harden the library and future dependencies against
forgetting to add is_always_stable to every specialization. We keep the
typedef in every specialization we ship to make sure that it's
impossible to use incorrectly, but this specialization of sorter_traits
makes it harder to get things accidentally wrong.