Fix sorting empty collections (issue #194)
merge_insertion_sort and slab_sort failed when sorting an empty
collection for the same reason: they tried to create an empty
fixed_size_list, which actually causes all knids of issues. The two
sorters don't attempt to creat such a list anymore, and an assertion was
added to the fixed_size_list constructor to ensure that it never
happens.