Replace some characters in doc before uploading to wiki
This is a first step to having a solution that works in GitHub Wiki,
Github repository browsing and Gollum. There is no perfect solution, but
using the relative .md paths works everywhere except on the GitHub wiki.
The simple solution is to get rid of them in the script that uploads the
documentation to the wiki.