Implement heap_sorter as a bottom-up heapsort
Update the heapsort algorithm to match the latest version of the code in
libc++. As a result heap_sort performs significantly fewer comparisons,
even though it doesn't make much of a speed difference for simple
scenarios.

cartesian_tree_sorter is also impacted, although the reduction in the
number of comparisons is mostly noticeable in its worst cas scenarios.