Add -Wno-inline
We almost never try to explicitly inline functions: it is always
implicit via templates or constexpr functions, and the rare functions
marked inline are marked as such to avoid ODR issues.

Somehow -Winline is massively triggered with MinGW-w64 (thousands of
lines of warnings when compiling the test suite), so disabling the
warning is probably the best thing to do.