CI: disable Clang builds on MacOS
In Clang 14, the CMake option LIBCXX_ENABLE_INCOMPLETE_FEATURES is
turned to OFF by default (it used to be ON) when building libc++, which
disables "incomplete" library features including <ranges>, which is
required to build cpp-sort 2.0.0.

We can't build the test suite or examples without <ranges>, so we will
have to wait for a subsequent revision with full support by default.