Fix slabsort with equivalent elements (#211)
slab_sort had a bug where it partitioned the collection stably around
the median, then recursed in each *half* of the collection instead of
recursing in the two partitions created by the partitioning algorithm.
It made no difference with unique elements since the partitions would
always correspond to both halves of the collection. However the story
was different when several elements compared equivalent to the median,
leading to the error encountered in #211.