Compite test suite with -Wno-inline with GNU frontends
Typically some flavours of MinGW produce around 13k lines of -Winline
warnings even though the library only uses inline sparringly and for ODR
reasons. There is no explicit "force inline" anywhere, so the warning is
just noise that clutters the compiler results and makes it impossible to
find other warnings in the logs.