Fix rendering of the logo
The SVG for the library's logo used text as-is but it referenced a font
that's not available on most computers. It happened to work on mine but
looked weird anymore else. The new SVG uses shape objects directly
instead of rendered text, and as such doesn't rely on the availability
on the font anymore.

I also took the opportunity to reduce it to the minimal required
components for it to display correctly, roughly halving its size.