iterable -> collection
The word iterable was a bit overloaded so I decided to split it into two
distinct power words for 2.x.y:
- Range when an iterable is required to model a range
- Collection otherwise

It is worth noting that "collection" does not imply that the type is
iterable, and some features are indeed not required to iterate over a
collection of elements provided that collection has a few other tricks
up its sleeve. It is a much looser concept than range and is not
formally defined.