Make verge_adapter work with bidirectional iterators
The only reason verge_adapter, unlike verge_sorter, did not work with
bidirectional iterators was the half-baked hacks using sized_iterator
which I couldn't reasonably impose on users, and I certainly did not
want to add more hacks to more sorters. Replacing size_iterator with
sized_range solved that issue.

This change includes a minr refactor to reuse a bit of split_adapter to
compute the iterator category of verge_adapter, which is done the same
way.