More accurate comparisons/projections wrt branchless traits
The classes comparisons_counter and projections_counter underlying those
metrics are now considered "probably branchless" when the function they
wrap is considered "probably branchless". This is sometimes a lie,
though what matters is that they give a result closer to the reality of
what happens when a sorter is not under analysis.

This also changes the behaviour of counting_adapter, which does not
really matter since it is deprecated.