Bug fix (see #230) + improvements to probe::exc
When a cycle started and ended with elements comparing equal, probe::exc
returned an incorrect result, which this commit fixes.

Additionally, this commit fixes the Exc(X) + 1 <= Ham(X) relation when X
is already sorted, and improves the space complexity of probe::exc by
getting rid of the 'sorted' vector in a similar way that what was
recently done for indirect_adapter.